Initial migration to centralised traits
