+++ src/Views/Filters/DateFilter.php
@@ -7,7 +7,7 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 
 class DateFilter extends Filter
 {
-    public string $viewPath = 'livewire-tables::components.tools.filters.date';
+    public string $view = 'livewire-tables::components.tools.filters.date';
 
     public function config(array $config = []): DateFilter
     {
@@ -47,8 +47,4 @@ class DateFilter extends Filter
         return $this->filterDefaultValue ?? null;
     }
 
-    public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
-    {
-        return view($this->getViewPath(), $this->getFilterDisplayData());
-    }
 }
