+++ src/Views/Filters/DateTimeFilter.php
@@ -7,7 +7,7 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 
 class DateTimeFilter extends Filter
 {
-    public string $viewPath = 'livewire-tables::components.tools.filters.datetime';
+    public string $view = 'livewire-tables::components.tools.filters.datetime';
 
     public function config(array $config = []): DateTimeFilter
     {
@@ -47,8 +47,4 @@ class DateTimeFilter extends Filter
         return $this->filterDefaultValue ?? null;
     }
 
-    public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
-    {
-        return view($this->getViewPath(), $this->getFilterDisplayData());
-    }
 }
