+++ src/Views/Filters/NumberRangeFilter.php
@@ -8,7 +8,7 @@ class NumberRangeFilter extends Filter
 {
     public array $options = [];
 
-    public string $viewPath = 'livewire-tables::components.tools.filters.number-range';
+    public string $view = 'livewire-tables::components.tools.filters.number-range';
 
     public function options(array $options = []): NumberRangeFilter
     {
@@ -86,10 +86,5 @@ class NumberRangeFilter extends Filter
         }
 
         return '';
-    }
-
-    public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
-    {
-        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }
