+++ src/Views/Filters/SelectFilter.php
@@ -8,7 +8,7 @@ class SelectFilter extends Filter
 {
     public array $options = [];
 
-    public string $viewPath = 'livewire-tables::components.tools.filters.select';
+    public string $view = 'livewire-tables::components.tools.filters.select';
 
     public function options(array $options = []): SelectFilter
     {
@@ -61,10 +61,5 @@ class SelectFilter extends Filter
     public function getFilterDefaultValue(): ?string
     {
         return $this->filterDefaultValue ?? null;
-    }
-
-    public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
-    {
-        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }
