+++ src/Views/Filters/TextFilter.php
@@ -6,7 +6,7 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 
 class TextFilter extends Filter
 {
-    public string $viewPath = 'livewire-tables::components.tools.filters.text-field';
+    public string $view = 'livewire-tables::components.tools.filters.text-field';
 
     public function validate(string $value): string|bool
     {
@@ -30,8 +30,4 @@ class TextFilter extends Filter
         return $this->filterDefaultValue ?? null;
     }
 
-    public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
-    {
-        return view($this->getViewPath(), $this->getFilterDisplayData());
-    }
 }
