+++ src/Views/Traits/Configuration/FilterConfiguration.php
@@ -139,11 +139,4 @@ trait FilterConfiguration
 
         return $this;
     }
-
-    public function setCustomView(string $customView): self
-    {
-        $this->viewPath = $customView;
-
-        return $this;
-    }
 }
