+++ src/Views/Traits/HasAttributes.php
@@ -3,13 +3,48 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Traits;
 
 use Closure;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\AttributesConfiguration;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\AttributesHelpers;
+use Illuminate\View\ComponentAttributeBag;
+use Rappasoft\LaravelLivewireTables\Views\{Column,Filter};
 
 trait HasAttributes
 {
-    use AttributesConfiguration,
-        AttributesHelpers;
-
     protected ?Closure $attributesCallback = null;
+
+    public function attributes(Closure $callback): self
+    {
+        $this->attributesCallback = $callback;
+
+        return $this;
+    }
+
+    public function getAttributesCallback(): ?Closure
+    {
+        return $this->attributesCallback;
+    }
+
+    public function hasAttributesCallback(): bool
+    {
+        return $this->attributesCallback !== null;
+    }
+
+    // TODO: Test
+    public function getAttributeBag($row)
+    {
+        return new ComponentAttributeBag($this->hasAttributesCallback() ? app()->call($this->getAttributesCallback(), ['row' => $row]) : []);
+    }
+
+    /**
+     * @param  array<mixed>  $attributes
+     * @return mixed
+     */
+    public function arrayToAttributes(array $attributes)
+    {
+        return implode(' ', array_map(function ($key) use ($attributes) {
+            if (is_bool($attributes[$key])) {
+                return $attributes[$key] ? $key : '';
+            }
+
+            return $key.'="'.$attributes[$key].'"';
+        }, array_keys($attributes)));
+    }
 }
